03-py-async-jobs/requirements.txt
fastapi
jetpack-io>=0.5.1.dev202201241643053138
uvicorn